change output
Change the output to a list of lists form with the second member of each list being a sspan object. Thus  [ [z^k1 = p,1 sspan(X1)], ...].

This new structure for the output makes it easier, I think, to collect the eigenspace for each eigenvector.

The sspan object is a simplifying span object. For now we call it sspan, not span, so that it doesn't interfere with the span object from the linear algebra package (that doesn't simplify).

Right now an sspan object doesn't do all that much--eventually, we'll add simplifications to it.  The vectors in a sspan object aren't guaranteed to be linearly independent--for a symbolic case, we can't always tell.