New function alt_eigen_fullratsubst
o New function alt_eigen_fullratsubst. This is pretty much fullratsubst, but there is a limit to the number of iterations.  I was having troubles with fullratsubst not finishing--I should look into this.

o I was using the wrong variable z instead of u in a few places--fixed that.

o Fixed up test results--all tests pass.